+++ climt/_core/__init__.py
@@ -2,7 +2,7 @@ from .util import (
     mass_to_volume_mixing_ratio,
     get_interface_values,
     bolton_q_sat, bolton_dqsat_dT)
-from .initialization import get_default_state
+from .initialization import get_default_state, quantity_descriptions
 from .arrays import (
     get_numpy_arrays_from_state,
     create_state_dict_for)
@@ -10,7 +10,7 @@ from .climt_components import ImplicitPrognostic
 
 __all__ = (
     mass_to_volume_mixing_ratio, get_default_state, get_numpy_arrays_from_state,
-    get_interface_values, create_state_dict_for,
+    get_interface_values, create_state_dict_for, quantity_descriptions,
     bolton_q_sat, bolton_dqsat_dT,
     ImplicitPrognostic
 )